[connector] Move utils to services
