+++ external-import/crowdstrike/src/crowdstrike_feeds_connector/actor/builder.py
@@ -4,7 +4,7 @@
 import logging
 from typing import List, Optional, Tuple
 
-from ..utils import (
+from crowdstrike_feeds_services.utils import (
     create_external_reference,
     create_intrusion_set,
     create_originates_from_relationships,
