+++ external-import/crowdstrike/src/crowdstrike_feeds_connector/indicator/importer.py
@@ -9,8 +9,8 @@ from .builder import (
     IndicatorBundleBuilder,
     IndicatorBundleBuilderConfig,
 )
-from ..utils import datetime_to_timestamp, timestamp_to_datetime
-from ..utils.report_fetcher import FetchedReport, ReportFetcher
+from crowdstrike_feeds_services.utils import datetime_to_timestamp, timestamp_to_datetime
+from crowdstrike_feeds_services.utils.report_fetcher import FetchedReport, ReportFetcher
 from crowdstrike_client.api.intel import Indicators, Reports
 from crowdstrike_client.api.models import Indicator
 from pycti.connector.opencti_connector_helper import (  # type: ignore  # noqa: E501
