+++ external-import/crowdstrike/src/crowdstrike_feeds_connector/rule/__init__.py
