+++ external-import/crowdstrike/src/crowdstrike_feeds_connector/rule/yara_master_importer.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""OpenCTI CrowdStrike Snort master importer module."""
+"""OpenCTI CrowdStrike YARA master importer module."""
 
 import itertools
 import zipfile
@@ -8,10 +8,10 @@ from io import BytesIO
 from typing import Any, Dict, List, Mapping, NamedTuple, Optional, Tuple
 
 from ..importer import BaseImporter
-from .snort_suricata_master_builder import SnortRuleBundleBuilder
-from ..utils import datetime_to_timestamp, timestamp_to_datetime
-from ..utils.report_fetcher import FetchedReport, ReportFetcher
-from ..utils.snort_parser import SnortParser, SnortRule
+from .yara_master_builder import YaraRuleBundleBuilder
+from crowdstrike_feeds_services.utils import datetime_to_timestamp, timestamp_to_datetime
+from crowdstrike_feeds_services.utils.report_fetcher import FetchedReport, ReportFetcher
+from crowdstrike_feeds_services.utils.yara_parser import YaraParser, YaraRule
 from crowdstrike_client.api.intel import Reports, Rules
 from crowdstrike_client.api.models.download import Download
 from pycti.connector.opencti_connector_helper import (  # type: ignore  # noqa: E501
@@ -22,19 +22,19 @@ from stix2 import Bundle, Identity, MarkingDefinition  # type: ignore
 from stix2.exceptions import STIXError  # type: ignore
 
 
-class SnortMaster(NamedTuple):
-    """Snort Master."""
+class YaraMaster(NamedTuple):
+    """YARA Master."""
 
-    rules: List[SnortRule]
+    rules: List[YaraRule]
     e_tag: Optional[str]
     last_modified: Optional[datetime]
 
 
-class SnortMasterImporter(BaseImporter):
-    """CrowdStrike Snort master importer."""
+class YaraMasterImporter(BaseImporter):
+    """CrowdStrike YARA master importer."""
 
-    _E_TAG = "snort_master_e_tag"
-    _LAST_MODIFIED = "snort_master_last_modified"
+    _E_TAG = "yara_master_e_tag"
+    _LAST_MODIFIED = "yara_master_last_modified"
 
     _KEY_ID = "id"
     _KEY_INDICATOR_PATTERN = "pattern"
@@ -50,7 +50,7 @@ class SnortMasterImporter(BaseImporter):
         report_status: int,
         report_type: str,
     ) -> None:
-        """Initialize CrowdStrike Snort master importer."""
+        """Initialize CrowdStrike YARA master importer."""
         super().__init__(helper, author, tlp_marking, update_existing_data)
 
         self.rules_api = rules_api
@@ -61,7 +61,7 @@ class SnortMasterImporter(BaseImporter):
 
     def run(self, state: Dict[str, Any]) -> Dict[str, Any]:
         """Run importer."""
-        self._info("Running Snort master importer with state: {0}...", state)
+        self._info("Running YARA master importer with state: {0}...", state)
 
         # Ignore the Etag, see the comment below.
         # e_tag = state.get(self._E_TAG)
@@ -71,59 +71,59 @@ class SnortMasterImporter(BaseImporter):
             last_modified = timestamp_to_datetime(last_modified)
 
         # XXX: Using Etag and Last-Modified results in HTTP 500.
-        # snort_master = self._fetch_snort_master(e_tag, last_modified)
-        snort_master = self._fetch_snort_master()
+        # yara_master = self._fetch_yara_master(e_tag, last_modified)
+        yara_master = self._fetch_yara_master()
 
-        latest_e_tag = snort_master.e_tag
-        latest_last_modified = snort_master.last_modified
+        latest_e_tag = yara_master.e_tag
+        latest_last_modified = yara_master.last_modified
 
         if (
             last_modified is not None
             and latest_last_modified is not None
             and last_modified >= latest_last_modified
         ):
-            self._info("Snort master not modified, skipping...")
+            self._info("YARA master not modified, skipping...")
             return state
 
-        snort_rules = snort_master.rules
-        snort_rule_count = len(snort_rules)
+        yara_rules = yara_master.rules
+        yara_rule_count = len(yara_rules)
 
         self._info(
-            "Snort master with {0} rules...",
-            snort_rule_count,
+            "YARA master with {0} rules...",
+            yara_rule_count,
         )
 
-        new_snort_rules = self._update_existing(snort_rules)
-        new_snort_rule_count = len(new_snort_rules)
+        new_yara_rules = self._update_existing(yara_rules)
+        new_yara_rule_count = len(new_yara_rules)
 
         self._info(
-            "{0} new Snort rules...",
-            new_snort_rule_count,
+            "{0} new YARA rules...",
+            new_yara_rule_count,
         )
 
-        grouped_snort_rules = self._group_snort_rules_by_report(new_snort_rules)
-        group_count = len(grouped_snort_rules)
+        grouped_yara_rules = self._group_yara_rules_by_report(new_yara_rules)
+        group_count = len(grouped_yara_rules)
 
         self._info(
-            "{0} Snort rule groups...",
+            "{0} YARA rule groups...",
             group_count,
         )
 
-        for group, rules in grouped_snort_rules:
-            self._info("Snort rule group: ({0}) {1}", len(rules), group)
+        for group, rules in grouped_yara_rules:
+            self._info("YARA rule group: ({0}) {1}", len(rules), group)
 
         failed_count = 0
 
-        for snort_rule_group in grouped_snort_rules:
-            failed = self._process_snort_rule_group(snort_rule_group)
+        for yara_rule_group in grouped_yara_rules:
+            failed = self._process_yara_rule_group(yara_rule_group)
             failed_count += failed
 
-        success_count = new_snort_rule_count - failed_count
+        success_count = new_yara_rule_count - failed_count
 
         self._info(
-            "Snort master importer completed (imported: {0}, total: {1}, e_tag: {2}, last_modified: {3})",  # noqa: E501
+            "YARA master importer completed (imported: {0}, total: {1}, e_tag: {2}, last_modified: {3})",  # noqa: E501
             success_count,
-            new_snort_rule_count,
+            new_yara_rule_count,
             latest_e_tag,
             latest_last_modified,
         )
@@ -144,52 +144,52 @@ class SnortMasterImporter(BaseImporter):
         self._info("Clearing report fetcher cache...")
         self.report_fetcher.clear_cache()
 
-    def _fetch_snort_master(
+    def _fetch_yara_master(
         self, e_tag: Optional[str] = None, last_modified: Optional[datetime] = None
-    ) -> SnortMaster:
-        download = self._fetch_latest_snort_master(
+    ) -> YaraMaster:
+        download = self._fetch_latest_yara_master(
             e_tag=e_tag, last_modified=last_modified
         )
-        return SnortMaster(
+        return YaraMaster(
             rules=self._parse_download(download),
             e_tag=download.e_tag,
             last_modified=download.last_modified,
         )
 
-    def _fetch_latest_snort_master(
+    def _fetch_latest_yara_master(
         self, e_tag: Optional[str] = None, last_modified: Optional[datetime] = None
     ) -> Download:
-        rule_set_type = "snort-suricata-master"
+        rule_set_type = "yara-master"
         return self.rules_api.get_latest_file(
             rule_set_type, e_tag=e_tag, last_modified=last_modified
         )
 
-    def _parse_download(self, download: Download) -> List[SnortRule]:
-        snort_str = self._unzip_content(download.content)
-        return self._parse_snort_rules(snort_str)
+    def _parse_download(self, download: Download) -> List[YaraRule]:
+        yara_str = self._unzip_content(download.content)
+        return self._parse_yara_rules(yara_str)
 
     @staticmethod
     def _unzip_content(compressed_content: BytesIO) -> str:
-        snort_master_filename = "crowdstrike_intel_snort.rules"
+        yara_master_filename = "crowdstrike_intel_yara.yara"
         with zipfile.ZipFile(compressed_content) as z:
-            with z.open(snort_master_filename) as snort_master:
-                return snort_master.read().decode("utf-8")
+            with z.open(yara_master_filename) as yara_master:
+                return yara_master.read().decode("utf-8")
 
     @staticmethod
-    def _parse_snort_rules(snort_rules: str) -> List[SnortRule]:
-        return SnortParser.parse(snort_rules)
+    def _parse_yara_rules(yara_rules: str) -> List[YaraRule]:
+        return YaraParser.parse(yara_rules)
 
-    def _update_existing(self, snort_rules: List[SnortRule]) -> List[SnortRule]:
-        """Update Snort rules if they already exists in the OpenCTI."""
-        new_snort_rules = []
+    def _update_existing(self, yara_rules: List[YaraRule]) -> List[YaraRule]:
+        """Update YARA rules if they already exists in the OpenCTI."""
+        new_yara_rules = []
 
         updated = 0
         not_updated = 0
 
-        for snort_rule in snort_rules:
-            rule_updated = self._try_updating(snort_rule)
+        for yara_rule in yara_rules:
+            rule_updated = self._try_updating(yara_rule)
             if rule_updated is None:
-                new_snort_rules.append(snort_rule)
+                new_yara_rules.append(yara_rule)
             else:
                 if rule_updated:
                     updated += 1
@@ -198,25 +198,25 @@ class SnortMasterImporter(BaseImporter):
 
         existing = updated + not_updated
 
-        self._info("Updated {0} of {1} existing Snort rules", updated, existing)
+        self._info("Updated {0} of {1} existing YARA rules", updated, existing)
 
-        return new_snort_rules
+        return new_yara_rules
 
-    def _try_updating(self, snort_rule: SnortRule) -> Optional[bool]:
-        """Try updating Snort rule if it already exists in the OpenCTI."""
-        name = snort_rule.name
+    def _try_updating(self, yara_rule: YaraRule) -> Optional[bool]:
+        """Try updating YARA rule if it already exists in the OpenCTI."""
+        name = yara_rule.name
 
         existing_rule = self._find_rule_by_name(name)
         if existing_rule is None:
             return None
 
-        return self._update_if_needed(snort_rule, existing_rule)
+        return self._update_if_needed(yara_rule, existing_rule)
 
     @staticmethod
-    def _group_snort_rules_by_report(
-        snort_rules: List[SnortRule],
-    ) -> List[Tuple[str, List[SnortRule]]]:
-        def _key_func(item: SnortRule) -> str:
+    def _group_yara_rules_by_report(
+        yara_rules: List[YaraRule],
+    ) -> List[Tuple[str, List[YaraRule]]]:
+        def _key_func(item: YaraRule) -> str:
             reports = item.reports
             if reports:
                 sorted_reports = sorted(reports)
@@ -224,40 +224,38 @@ class SnortMasterImporter(BaseImporter):
             return ""
 
         groups = []
-        sorted_snort_rules = sorted(snort_rules, key=_key_func)
-        for key, group in itertools.groupby(sorted_snort_rules, key=_key_func):
+        sorted_yara_rules = sorted(yara_rules, key=_key_func)
+        for key, group in itertools.groupby(sorted_yara_rules, key=_key_func):
             groups.append((key, list(group)))
         return groups
 
-    def _process_snort_rule_group(
-        self, snort_rule_group: Tuple[str, List[SnortRule]]
+    def _process_yara_rule_group(
+        self, yara_rule_group: Tuple[str, List[YaraRule]]
     ) -> int:
-        group = snort_rule_group[0]
-        self._info("Processing Snort rule group '{0}'...", group)
+        group = yara_rule_group[0]
+        self._info("Processing YARA rule group '{0}'...", group)
 
-        snort_rules = snort_rule_group[1]
-        total_count = len(snort_rules)
+        yara_rules = yara_rule_group[1]
+        total_count = len(yara_rules)
 
         failed_count = 0
 
-        for snort_rule in snort_rules:
-            fetched_reports = self._get_reports_by_code(snort_rule.reports)
+        for yara_rule in yara_rules:
+            fetched_reports = self._get_reports_by_code(yara_rule.reports)
 
-            snort_rule_bundle = self._create_snort_rule_bundle(
-                snort_rule, fetched_reports
-            )
-            if snort_rule_bundle is None:
+            yara_rule_bundle = self._create_yara_rule_bundle(yara_rule, fetched_reports)
+            if yara_rule_bundle is None:
                 failed_count += 1
 
-            # with open(f"snort_rule_bundle_{snort_rule.name}.json", "w") as f:
-            #     f.write(snort_rule_bundle.serialize(pretty=True))
+            # with open(f"yara_rule_bundle_{yara_rule.name}.json", "w") as f:
+            #     f.write(yara_rule_bundle.serialize(pretty=True))
 
-            self._send_bundle(snort_rule_bundle)
+            self._send_bundle(yara_rule_bundle)
 
         success_count = total_count - failed_count
 
         self._info(
-            "Completed processing Snort rule group '{0}' (imported: {1}, total: {2})",
+            "Completed processing YARA rule group '{0}' (imported: {1}, total: {2})",
             group,
             success_count,
             total_count,
@@ -266,7 +264,7 @@ class SnortMasterImporter(BaseImporter):
         return failed_count
 
     def _update_if_needed(
-        self, new_rule: SnortRule, existing_rule: Tuple[str, SnortRule]
+        self, new_rule: YaraRule, existing_rule: Tuple[str, YaraRule]
     ) -> bool:
         new_rule_name = new_rule.name
         indicator_id, current_rule = existing_rule
@@ -281,7 +279,7 @@ class SnortMasterImporter(BaseImporter):
             self._info("Not updating rule '{0}' ({1})", new_rule_name, indicator_id)
             return False
 
-    def _find_rule_by_name(self, name: str) -> Optional[Tuple[str, SnortRule]]:
+    def _find_rule_by_name(self, name: str) -> Optional[Tuple[str, YaraRule]]:
         indicator = self._fetch_indicator_by_name(name)
         if indicator is None:
             return None
@@ -296,15 +294,15 @@ class SnortMasterImporter(BaseImporter):
             self._error("Indicator '{0}' without pattern", name)
             return None
 
-        rules = SnortParser.parse(indicator_pattern)
+        rules = YaraParser.parse(indicator_pattern)
 
         if not rules:
-            self._error("Indicator '{0}' pattern without Snort rules", name)
+            self._error("Indicator '{0}' pattern without YARA rules", name)
             return None
 
         if len(rules) > 1:
             self._error(
-                "Indicator '{0}' pattern contains more than one Snort rules", name
+                "Indicator '{0}' pattern contains more than one YARA rules", name
             )
             return None
 
@@ -319,10 +317,10 @@ class SnortMasterImporter(BaseImporter):
         }
         return self.helper.api.indicator.read(filters=filters)
 
-    def _needs_updating(self, current_rule: SnortRule, new_rule: SnortRule) -> bool:
+    def _needs_updating(self, current_rule: YaraRule, new_rule: YaraRule) -> bool:
         if current_rule.name != new_rule.name:
             self._error(
-                "Current ({0}) and new ({1}) Snort rules names do no match",
+                "Current ({0}) and new ({1}) YARA rules names do no match",
                 current_rule.name,
                 new_rule.name,
             )
@@ -357,8 +355,8 @@ class SnortMasterImporter(BaseImporter):
             self._error("Failed to fetch reports {0}: {1}", codes, e)
             return []
 
-    def _create_snort_rule_bundle(
-        self, rule: SnortRule, reports: List[FetchedReport]
+    def _create_yara_rule_bundle(
+        self, rule: YaraRule, reports: List[FetchedReport]
     ) -> Optional[Bundle]:
         author = self.author
         source_name = self._source_name()
@@ -367,7 +365,7 @@ class SnortMasterImporter(BaseImporter):
         report_status = self.report_status
         report_type = self.report_type
 
-        bundle_builder = SnortRuleBundleBuilder(
+        bundle_builder = YaraRuleBundleBuilder(
             rule,
             author,
             source_name,
@@ -382,7 +380,7 @@ class SnortMasterImporter(BaseImporter):
             return bundle_builder.build()
         except STIXError as e:
             self._error(
-                "Failed to build Snort rule bundle for '{0}': {1}",
+                "Failed to build YARA rule bundle for '{0}': {1}",
                 rule.name,
                 e,
             )
