+++ external-import/crowdstrike/src/crowdstrike_feeds_services/client/endpoints.py
@@ -1,0 +1,4 @@
+"""
+API URL Constants
+"""
+
