+++ external-import/crowdstrike/src/crowdstrike_feeds_services/utils/observables.py
