+++ external-import/crowdstrike/src/crowdstrike_feeds_services/utils/snort_parser.py
