+++ external-import/crowdstrike/src/crowdstrike_feeds_services/utils/yara_parser.py
