+++ external-import/crowdstrike/src/requirements.txt
@@ -1,4 +1,5 @@
 -e git+https://github.com/certeu/crowdstrike-client#egg=crowdstrike-client
 pycti==6.1.12
 lxml==5.2.2
+crowdstrike-falconpy==1.4.3
 pydantic==1.10.17
